Checkpoint commit.
-- Redesign the 'Comments' pane of the Create Entry page to accomodate the new
comment-closing options more logically and shuffle things around a bit to
make the available options clearer. (Hopefully much clearer.)

-- Implement the option to disable top-level replies to an entry, but still
allow replies-to-replies (requested in http://dw-suggestions.dreamwidth.org/442319.html ).

-- ...which necessitated me rethinking how I was going to handle the logprop,
so there's a new logprop for governing comment-closing, and the one I had been
using is being removed.

Of note: Redesigning the 'Comments' pane also resulted in me realizing
that the existing "disable comments entirely and don't show the existing
comments" (handled by the 'opt_nocomments' logprop) logically belonged in
the same dropdown as the other close-comments options. However, that logprop
is set and called in a *lot* of places (especially in crossposting), and
changing it opened up the possibility of way too many heisenbugs. So, the
changes to the logic of cgi-bin/DW/Controller/Entry.pm are a wee bit hacky,
but it's the best way I could think of to accomodate the legacy logprop.

This has been very thoroughly tested in just about every permutation I could
think of, but more testing is always welcome!

Still to do:
* add the appropriate text properties into non-TR-based layouts
* add a 'close comments after $time' option on creating an entry
* add account-wide preference for closing comments on a timed basis
* work out how I'm going to actually close the comments on time.